refactor: move flag to a different shared module
